working on structure changes
