+++ include/server.hpp
@@ -1,0 +1,10 @@
+#ifndef __SERVER_HEADER__
+#define __SERVER_HEADER__
+
+class Server
+{
+public:
+    void startProgram();
+};
+
+#endif
